Add more detailed BLS data and deliver via API
We need to have BLS data by category, region and pay range.
Since changing region and pay range are being considered as user
choices, the BLS data will need to be available on demand with more
specific API calls or from a BLS json object that has all the choices
available.

Region will be available from a new schoo.region value.
Income will come either from program.salary,
or school.medianAnnualPay as a fallback.
With those two values in hand, the appropriate BLS expense data could be
plucked from a bls-data API call.
