+++ paying_for_college/tests/test_views.py
@@ -8,7 +8,7 @@ from django.test import RequestFactory
 from django.http import HttpResponse, HttpRequest
 from django.test import Client
 from django.core.urlresolvers import reverse
-from paying_for_college.views import get_school, get_region, validate_oid
+from paying_for_college.views import get_school, validate_oid
 from paying_for_college.views import get_program, get_program_length
 from paying_for_college.views import Feedback, EmailLink
 from paying_for_college.views import SchoolRepresentation, school_search_api
@@ -64,14 +64,6 @@ class TestViews(django.test.TestCase):
         program.level = '3'
         test4 = get_program_length(program=program, school='')
         self.assertTrue(test4 == 4)
-
-    def test_get_region(self):
-        school = School(school_id='123456', state='NE')
-        self.assertTrue(get_region(school) == 'MW')
-
-    def test_get_region_failure(self):
-        school = School(school_id='123456', state='')
-        self.assertTrue(get_region(school) == '')
 
     def test_landing_page_views(self):
         for url_name in self.landing_page_views:
@@ -213,9 +205,9 @@ class OfferTest(django.test.TestCase):
         resp6 = client.get(url+missing_school_id)
         self.assertTrue("doesn't contain a school" in resp6.content)
         self.assertTrue(resp6.status_code == 400)
-        resp7 = client.get(url+puerto_rico)
-        self.assertTrue("nationalFood" in resp7.content)
-        self.assertTrue(resp7.status_code == 200)
+        # resp7 = client.get(url+puerto_rico)
+        # self.assertTrue("nationalFood" in resp7.content)
+        # self.assertTrue(resp7.status_code == 200)
         resp8 = client.get(url+bad_oid)
         self.assertTrue("illegal characters" in resp8.content)
         self.assertTrue(resp8.status_code == 400)
@@ -257,9 +249,9 @@ class APITests(django.test.TestCase):
         resp2 = client.get(url2)
         self.assertTrue('No school' in resp2.content)
         self.assertTrue(resp2.status_code == 400)
-        url3 = reverse('disclosures:national-stats-json', args=['243197'])
-        resp3 = client.get(url3)
-        self.assertTrue('nationalFood' in resp3.content)
+        # url3 = reverse('disclosures:national-stats-json', args=['243197'])
+        # resp3 = client.get(url3)
+        # self.assertTrue('nationalFood' in resp3.content)
 
     # /paying-for-college/understanding-financial-aid-offers/api/program/408039-981/
     def test_program_json(self):
